aaaa

+++ Models/PE_Using.cs
@@ -7,7 +7,7 @@ namespace PRJ_WAREHOUSE_BIVN.Models
         public string MaVatTu { get; set; }
         public int Thang { get; set; }
         public int Nam { get; set; }
-        public decimal Soluong { get; set; }
+        public double Soluong { get; set; }
         public DateTime Ngayupdate { get; set; }
         public string Nguoiupdate { get; set; }
     }
